Add make.hint.report
make.hint.report creates an Rmd file from a problem set that shows for every chunk the sample solution, the given task code and the hint() message that would be shown if the task code would not have changed. This reports helps to see if hints seem well. Helpful for custom hints as well as improvement of the automatic hint functions.